refactor: transition to command-based renderer
