+++ AtmosphericEngine/include/Atmospheric/graphics_server.hpp
@@ -13,11 +13,7 @@
 #include <glm/vec4.hpp>
 #include <unordered_map>
 
-
-enum class RenderPath {
-    Forward,
-    Deferred
-};
+enum class DrawMode { Static, Dynamic, Stream };
 
 struct CanvasVertex {
     glm::vec2 position;
@@ -45,12 +41,6 @@ struct InstanceData {
     glm::mat4 modelMatrix;
 };
 
-struct RenderTargetProps {
-    int width = INIT_FRAMEBUFFER_WIDTH;
-    int height = INIT_FRAMEBUFFER_HEIGHT;
-    int numSamples = MSAA_NUM_SAMPLES;
-};
-
 // class Pipeline;
 
 // class ShaderProgram;
@@ -58,6 +48,8 @@ struct RenderTargetProps {
 // class Texture;
 
 // class Material;
+
+class Renderer;
 
 class GraphicsServer : public Server {
 private:
@@ -74,6 +66,17 @@ public:
     std::vector<LightComponent*> pointLights;
     std::vector<CameraComponent*> cameras;
 
+    std::unordered_map<Mesh*, std::vector<InstanceData>> _meshInstanceMap;
+    std::vector<DebugVertex> debugLines;
+    std::vector<CanvasVertex> canvasDrawList;
+    int _debugLineCount = 0;
+    int _canvasQuadCount = 0;
+
+    Mesh* debugLineMesh = nullptr;
+    Mesh* canvasMesh = nullptr;
+    ShaderProgram* debugShader = nullptr;
+    ShaderProgram* canvasShader = nullptr;
+
     GraphicsServer();
     ~GraphicsServer();
 
@@ -83,6 +86,11 @@ public:
 
     void Reset();
     void Render(float dt);
+
+    void PushDebugLine(DebugVertex from, DebugVertex to) {
+        debugLines.push_back(from);
+        debugLines.push_back(to);
+    }
 
     CameraComponent* GetMainCamera() const {
         if (cameras.size() > 0) {
@@ -104,10 +112,18 @@ public:
     ShaderProgram* GetShaderByID(uint32_t id) const;
     Mesh* GetMesh(const std::string& name) const;
 
-    void CheckFramebufferStatus(const std::string& prefix);
-    void CheckErrors(const std::string& prefix);
+    MeshComponent* RegisterMesh(MeshComponent* mesh);
+    CameraComponent* RegisterCamera(CameraComponent* camera);
+    LightComponent* RegisterLight(LightComponent* light);
+    SpriteComponent* RegisterSprite(SpriteComponent* sprite);
 
-    void PushDebugLine(DebugVertex from, DebugVertex to);
+    Renderer* renderer = nullptr;
+
+private:
+    CameraComponent* defaultCamera = nullptr;
+    LightComponent* defaultLight = nullptr;
+
+    static constexpr int MAX_CANVAS_TEXTURES = 32;
 
     void PushCanvasQuad(
       float x,
@@ -135,104 +151,4 @@ public:
       const glm::vec2& tilesetSize,
       const glm::vec2& tileIndex
     );
-
-    void EnableWireframe(bool enable = true) {
-        wireframeEnabled = enable;
-    }
-
-    void EnablePostProcess(bool enable = true) {
-        postProcessEnabled = enable;
-    }
-
-    void SetCapability(const GLenum& cap, bool enable = true) {
-        if (enable) {
-            glEnable(GL_CULL_FACE);
-        } else {
-            glDisable(GL_CULL_FACE);
-        }
-    }
-
-
-    MeshComponent* RegisterMesh(MeshComponent* mesh);
-    CameraComponent* RegisterCamera(CameraComponent* camera);
-    LightComponent* RegisterLight(LightComponent* light);
-    SpriteComponent* RegisterSprite(SpriteComponent* sprite);
-
-    // void ApplyMaterial(const Material& material);
-
-    // void ApplyPipeline(const Pipeline& pipeline);
-
-    // void ApplyCamera(const Camera& camera);
-
-    // void ApplyTransform(const glm::mat4 transform);
-
-    // void Draw(const std::shared_ptr<Mesh> mesh);
-
-    void SetRenderPath(RenderPath path) {
-        _currRenderPath = path;
-    }
-    RenderPath GetRenderPath() const {
-        return _currRenderPath;
-    }
-
-private:
-    RenderPath _currRenderPath = RenderPath::Forward;
-
-    GLuint debugVAO, debugVBO;
-    GLuint canvasVAO, canvasVBO;
-    GLuint screenVAO, screenVBO;
-
-    std::array<GLuint, MAX_UNI_LIGHTS> uniShadowMaps;
-    std::array<GLuint, MAX_OMNI_LIGHTS> omniShadowMaps;
-    GLuint sceneColorTexture, sceneDepthTexture, sceneStencilTexture;
-    GLuint msaaResolveTexture;
-
-
-    GLuint shadowFBO, sceneFBO, msaaResolveFBO;
-    GLuint finalFBO = 0;
-    struct GBuffer {
-        GLuint id;
-        GLuint positionRT;
-        GLuint normalRT;
-        GLuint albedoRT;
-        GLuint materialRT;
-        GLuint depthRT;
-    } gBuffer;
-
-    std::unordered_map<Mesh*, std::vector<InstanceData>> _meshInstanceMap;
-    // std::unordered_map<std::pair<Mesh*, Material*>, std::vector<InstanceData>> groupedObjects;
-
-    std::vector<DebugVertex> debugLines;
-    std::vector<CanvasVertex> canvasDrawList;
-
-    glm::vec4 clearColor = glm::vec4(0.15f, 0.183f, 0.2f, 1.0f);
-    bool postProcessEnabled = false;
-    bool wireframeEnabled = false;
-
-    CameraComponent* defaultCamera = nullptr;
-    LightComponent* defaultLight = nullptr;
-
-    int auxLightCount = 0;
-    int _canvasQuadCount = 0;
-    int _debugLineCount = 0;
-
-    void CreateFBOs();
-    void DestroyFBOs();
-    void CreateRTs(const RenderTargetProps& props);
-    void DestroyRTs();
-
-    void CreateCanvasVAO();
-    void CreateScreenVAO();
-    void CreateDebugVAO();
-
-    void ShadowPass(float dt);
-    void ForwardPass(float dt);
-    void MSAAResolvePass(float dt);
-    void CanvasPass(float dt);
-    void PostProcessPass(float dt);
-
-    void GeometryPass(float dt);
-    void LightingPass(float dt);
-
-    static constexpr int MAX_CANVAS_TEXTURES = 32;
 };
